requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
dash
dash-bootstrap-components
pandas
sqlalchemy
psycopg2-binary
plotly
python-dotenv
requests